+++ packages/core/stories/components/typography/CounterController.tsx
@@ -1,17 +1,28 @@
 import { useState } from 'react'
-import { Chart, Surface, View, Counter, CounterProps } from '../../../src'
+import { Chart, Surface, View, Counter, CounterProps, CounterValueProps } from '../../../src'
 
 const randomUniformValue = (min: number, max: number) => {
     return min + Math.random() * (max - min)
 }
 
+export const CustomCounterValue = ({ style, className, children }: CounterValueProps) => {
+    return (
+        <text style={style} className={className}>
+            {children}
+            <tspan style={{ fontSize: 14, dominantBaseline: 'text-before-edge' }}> %</tspan>
+        </text>
+    )
+}
+
 export const CounterController = ({
     position = [0, 0],
+    nDecimalPlaces = 0,
+    component,
     format = (v: number) => String(v),
-}: Pick<CounterProps, 'position' | 'format'>) => {
-    const [value, setValue] = useState(Math.round(randomUniformValue(0, 100)))
+}: Pick<CounterProps, 'position' | 'format' | 'nDecimalPlaces' | 'component'>) => {
+    const [value, setValue] = useState(randomUniformValue(0, 100))
     const updateValue = () => {
-        setValue(Math.round(randomUniformValue(0, 100)))
+        setValue(randomUniformValue(0, 100))
     }
     return (
         <>
@@ -25,7 +36,13 @@ export const CounterController = ({
             >
                 <Surface variant={'inner'} />
                 <View>
-                    <Counter position={position} format={format} style={{ fontSize: 24 }}>
+                    <Counter
+                        position={position}
+                        format={format}
+                        style={{ fontSize: 32 }}
+                        nDecimalPlaces={nDecimalPlaces}
+                        component={component}
+                    >
                         {value}
                     </Counter>
                 </View>