+++ packages/core/tests/typography/Counter.test.tsx
@@ -1,10 +1,10 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
-import { Chart, Counter } from '../../src'
+import { Chart, Counter, TextContentProps } from '../../src'
 import { chartProps } from '../props'
 import { useState } from 'react'
 
 describe('Counter', () => {
-    it('creates a default counter component', () => {
+    it('creates a default counter', () => {
         render(
             <Chart {...chartProps}>
                 <Counter>50</Counter>
@@ -13,6 +13,18 @@ describe('Counter', () => {
         expect(screen.getByRole('counter')).toBeDefined()
         const result = screen.getByRole('chart-content').querySelectorAll('text')
         expect(result).toHaveLength(1)
+        expect(result[0].getAttribute('class')).toContain('counter')
+    })
+
+    it('creates a counter without role', () => {
+        render(
+            <Chart {...chartProps}>
+                <Counter setRole={false}>50</Counter>
+            </Chart>
+        )
+        const result = screen.getByRole('chart-content').querySelector('text')
+        expect(result?.getAttribute('role')).toBeNull()
+        expect(result?.getAttribute('class')).toContain('counter')
     })
 
     it('updates values', async () => {
@@ -46,5 +58,34 @@ describe('Counter', () => {
         await waitFor(() => {
             expect(counter.textContent).toEqual('200')
         })
+    })
+
+    it('rounds values to a fixed number of decimal places', () => {
+        render(
+            <Chart {...chartProps}>
+                <Counter nDecimalPlaces={2}>50.123123</Counter>
+            </Chart>
+        )
+        const result = screen.getByRole('chart-content').querySelector('text')
+        expect(result?.textContent).toBe('50.12')
+    })
+
+    it('displays values with custom component', () => {
+        const CustomValue = ({ children, className }: TextContentProps) => (
+            <text className={className}>
+                <tspan>abc</tspan> {children} <tspan>xyz</tspan>
+            </text>
+        )
+        render(
+            <Chart {...chartProps}>
+                <Counter component={CustomValue}>50</Counter>
+            </Chart>
+        )
+        // text element should be annotated as a counter
+        const text = screen.getByRole('chart-content').querySelector('text')
+        expect(text?.closest('g')?.getAttribute('role')).toContain('counter')
+        expect(text?.getAttribute('class')).toContain('counter')
+        // there should be two tspan elements
+        expect(screen.getByRole('chart-content').querySelectorAll('tspan')).toHaveLength(2)
     })
 })